crashdump: fix ram_top overflow in compressed dump
gd->ram_top is ulong (32-bit on 32-bit arch). For DDR at 0x80000000
with 2GB, it wraps to 0x0 breaking the compression buffer check.

Use uint64_t type casting during ram_top calculation to avoid
address concatination.

+++ arch/arm/mach-ipq/cmd_crashdump.c
@@ -3322,8 +3322,9 @@ static void ipq_dump_func(crashdump_config_t *dump_config, uint8_t debug)
 	bool skip_crashdump = 1;
 
 	dump_config->debug = debug;
-	dump_config->ram_top = ((gd->ram_top & BIT(0)) ? (gd->ram_top + 1) :
-					gd->ram_top);
+	dump_config->ram_top = ((gd->ram_top & BIT(0)) ? ((uint64_t)gd->ram_top + 1) :
+				gd->ram_top);
+
 	parse_crashdump_config(dump_config);
 	if (!dump_config->force_collect_dump) {
 		etime = get_timer(0) + (10 * CONFIG_SYS_HZ);
